Polished a bit exception-related files.

+++ PyGMO/boost_python_container_conversions.h
@@ -179,7 +179,7 @@
         bool is_range = PyRange_Check(obj_ptr);
         std::size_t i=0;
         if (!all_elements_convertible(obj_iter, is_range, i)) return 0;
-        if (!is_range) assert(i == obj_size);
+        if (!is_range) assert(i == (size_t)obj_size);
       }
       return obj_ptr;
     }
